changed (long int) to GLint so compilation doesn't choke on recent headers...
git-svn-id: https://pd-gem.svn.sourceforge.net/svnroot/pd-gem/trunk/Gem@2226 49b2a0ca-c541-4d49-a7ef-d785e4371512

+++ src/Manips/glsl_program.cpp
@@ -426,7 +426,7 @@ void glsl_program :: LinkProgram()
   getVariables();
 # ifdef __APPLE__
   // call API to check if linked program is running on hardware or in software emulation
-  long int vertexGPUProcessing, fragmentGPUProcessing;
+  GLint vertexGPUProcessing, fragmentGPUProcessing;
   CGLGetParameter (CGLGetCurrentContext(), kCGLCPGPUVertexProcessing, &vertexGPUProcessing);
   CGLGetParameter (CGLGetCurrentContext(), kCGLCPGPUFragmentProcessing, &fragmentGPUProcessing);
